+++ src/main/java/edu/eci/pdsw/sampleprj/dao/mybatis/mappers/ClienteMapper.java
@@ -11,7 +11,7 @@ import org.apache.ibatis.annotations.Param;
  */
 public interface ClienteMapper {
     public void insertarCliente(Cliente ct);
-	public Cliente consultarCliente(@Param("idcli") int id);
+	public Cliente consultarCliente(@Param("idcli") long id);
     
     /**
      * Registrar un nuevo item rentado asociado al cliente identificado
